requirements.txt
requests>=2.31.0
python-frontmatter>=1.0.0
PyYAML>=6.0

